% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_count.R
\name{lib_count}
\alias{lib_count}
\title{Report and count the functions used}
\usage{
lib_count(path)
}
\arguments{
\item{path}{a .R file}
}
\value{
dataframe
}
\description{
Input a R script to find and count all the functions used in each respective package
}
