% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_lines.R
\name{lib_lines}
\alias{lib_lines}
\title{Document the dependancies}
\usage{
lib_lines(path)
}
\arguments{
\item{script}{a .R file}
}
\value{
markdown
}
\description{
Input a R script to prepare a markdown file documenting the dependancies. These can be included in a project
Readme file.
}
