% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_search.R
\name{lib_search}
\alias{lib_search}
\title{Find dependant packages}
\usage{
lib_search(script)
}
\arguments{
\item{script}{a .R file}
}
\value{
vector
}
\description{
Input a R script to find all the dependant packages used and create a character vector with all of their names.
}
\examples{
lib_search("script.R")
}
