% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_writeme.R
\name{lib_writeme}
\alias{lib_writeme}
\title{Document the dependancies}
\usage{
lib_writeme(script)
}
\arguments{
\item{script}{a .R file}
}
\value{
markdown
}
\description{
Input a R script to prepare a markdown file documenting the dependancies. These can be included in a project
Readme file.
}
\examples{
lib_writeme("script.R")
}
