% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librely-package.R
\docType{package}
\name{librely}
\alias{librely}
\alias{librely-package}
\title{librely: Meta analysis and documentation of R scripts}
\description{
librely provides a collection of functions to aid in the workflow of a data science
or software development project. The functions find and report the dependancies and
functions that were used in a R script. These dependancies can also be neatly
prepared as a text document to be included in a project repository.
}
